FM - Villain Actions
If Villain Actions are detected, sets the Legendary Action Uses fields to 3 

And some big tweaks to actionTitleRegex - it took me quite a while to get to something that was working for several of the different roadblocks I came up against, hopefully these changes don't impact other sources (only tested with FM 

`#actionTitleRegex = /^((?!.*has three villain actions)(Action \d+: )?([A-Z][\w\d\-+,;']?[\s\-]?)((of|and|the|from|in|at|on|with|to|by|into)\s)?([\w\d\-+,;'’]+\s?){0,5}(\((?!spell save)[^)]+\))?)[.!?]/;`

`(?!.*has three villain actions)` - rejects matches with the text "has three villain actions" this was necessary due to the increased word allotment below (it was creating a new feature with the title "X has three villain actions", which broke the Legendary Action Uses function)  
`(Action \d+: )?` - Optional match for villain actions, using this instead of adding `:` to the allowed punctuation saves one word allotment  
`([A-Z][\w\d\-+,;']?[\s\-]?)` - changed the `+` in between the `[ ]` groups to a `?` because a couple VAs start with "I " and they werent being captured.  
`[\w\d\-+,;'’]` - added `’` (weird apostrophe) that is used in a few VAs  
`{0,5}` - increase word allotment from 3 to 5 to account for all VAs (Queen Bargnot has the longest one) I think this could open the possibility of some unintended consequences for other statblocks?  
`[.!?]` - added `?` as a few VAs use it to end the title